requirements.txt
# requirements.txt
cp-mgmt-api-sdk
python-dotenv
Jinja2
smtplib
